requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
Pillow
pdfplumber
pytesseract
networkx
